% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{annotate_corner}
\alias{annotate_corner}
\title{Write text to one of four corners of a plot}
\usage{
annotate_corner(text, pos = "tl", geom = c("text", "label"), ...)
}
\arguments{
\item{text}{character string}

\item{pos}{character of \code{'tl'}, \code{'tr'}, \code{'bl'}, \code{'br'} to
indicate position}

\item{geom}{"text" or "label" geoms}

\item{...}{additional options passed to \link[ggplot2:geom_label]{ggplot2::geom_label}}
}
\value{
ggplot layer
}
\description{
Write text to one of four corners of a plot
}
\details{
If multiple panels, will write to all panels.
}
\examples{
example_plot()+annotate_corner('Hi.')

example_plot()+annotate_corner('Hi.', "bottomright")
}
