% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_normal_sample}
\alias{data_normal_sample}
\title{Testing dataset of grouped Normal distributions}
\usage{
data_normal_sample(mu = c(-0.5, 4), n = 500L, sd_range = c(0.6, 1.4),
  seed = 19850519)
}
\arguments{
\item{mu}{Means for each group. A \link{numeric} vector with the length
corresponding to the number of groups.}

\item{n}{Number of observations for each group. Length 1 \link{integer}.}

\item{sd_range}{The min and max to use for standard deviations. Length 2}

\item{seed}{A seed value to generate the same sample. \link{numeric} vector.}
}
\value{
A \link{data.frame} with the following variables: \code{grp}, \code{cond}, \code{value}.
}
\description{
Testing dataset of grouped Normal distributions
}
\examples{
data_normal_sample(0, 100)
}
