% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdistribute-package.R
\docType{package}
\name{ggdistribute-package}
\alias{ggdistribute-package}
\title{A 'ggplot2' Extension for Plotting Unimodal Distributions}
\description{
The \code{ggdistribute} package is an extension for plotting posterior or other
types of unimodal distributions that require overlaying information about a
distribution's intervals. It makes use of the \code{ggproto} system to extend
\code{ggplot2}, providing additional "geoms", "stats", and "positions." The
extensions integrate with existing \code{ggplot2} layer elements.
}
\details{
Displaying the distributions relies heavily on stacking distributions using
position_spread, which may not align with other geoms.
}
\seealso{
\link[ggplot2:geom_density]{ggplot2::geom_density}, \link[ggplot2:position_dodge]{ggplot2::position_dodge}

See \code{help(package = "ggdistribute")} for a list of functions.

View vignettes with \code{browseVignettes(package ="ggdistribute")}.
}
