% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{label_plot}
\alias{label_plot}
\title{Add labels to existing plot}
\usage{
label_plot(labels, x, y, g = list(fontsize = 14, fontface = "bold"), ...)
}
\arguments{
\item{labels}{character vector of labels to use}

\item{x}{horz positions of items in \code{labels}}

\item{y}{vert positions of items in \code{labels}}

\item{g}{list of options passed to \link[grid:gpar]{grid::gpar}}

\item{...}{optional args passed to \link[grid:grid.text]{grid::grid.text}}
}
\value{
NULL. prints to current graphics device.
}
\description{
Uses a normalized coordinate system to add text anywhere on the current plot.
}
\examples{
example_plot()
label_plot(c('a label', 'another one'), c(.1, .9), c(.95, .1))

# use extra options from grid::grid.text
label_plot('last one', 0.5, 0.5, just='center')
}
\seealso{
\link[grid:grid.text]{grid::grid.text}, \link[grid:gpar]{grid::gpar}
}
