% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghelpers.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Print a small example plot with geom_posterior}
\usage{
posterior_plot(data, x, y = "..count..", ...)
}
\arguments{
\item{data}{A dataset to use, called at the top layer within \code{ggplot}.}

\item{x}{A character string of the \code{x} axis variable name (e.g., values
making up the distribution).}

\item{y}{A grouping variable for generating groups of distributions. Defaults
to \code{..count..} for no groups and displays the density as counts of the
number of samples for the value of \code{x}.}

\item{...}{Additional arguments passed to \code{\link[=geom_posterior]{geom_posterior()}}.}
}
\value{
Object of class \code{gg}, \code{ggplot}.
}
\description{
Print a small example plot with geom_posterior
}
\examples{
# Generate a basic example plot if no data is specified.
posterior_plot()
}
