% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghelpers.R
\name{scale_add}
\alias{scale_add}
\title{scale and add}
\usage{
scale_add(base_size, amount = 1, adj = 0)
}
\arguments{
\item{base_size}{start value}

\item{amount}{multiple by}

\item{adj}{add after}
}
\value{
numeric
}
\description{
scale and add
}
