% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{show_colors}
\alias{show_colors}
\title{plot and show hex values of colors}
\usage{
show_colors(colors, show.legend = TRUE, ncols = NULL, alpha = NA)
}
\arguments{
\item{colors}{character vector of hex value colors}

\item{show.legend}{show the legend with hex values (logical)}

\item{ncols}{number of columns in the plot}
}
\value{
A plot with the index of the color in the tile
}
\description{
plot and show hex values of colors
}
\examples{
show_colors(mejr_palette())
show_colors(topo.colors(25))
show_colors(cm.colors(64), FALSE)
show_colors(viridisLite::viridis(15), alpha = .8)
}
