% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sre}
\alias{sre}
\alias{sre_data}
\title{Raw SRE dataset}
\format{Dataset is an object of class \link[tibble:tibble]{tibble::tibble}.}
\usage{
sre

sre_data(n = 1000, seed = 19850519)
}
\arguments{
\item{n}{number of samplers per effect and contrast}

\item{seed}{set.seed number}
}
\value{
A dataset of \link[tibble:tibble]{tibble::tibble}.
}
\description{
Raw SRE dataset
}
\section{Functions}{
\itemize{
\item \code{sre_data}: create samples from sre data
}}

\examples{
sre

sre_data()
}
\seealso{
\code{\link[=sre_data]{sre_data()}}
}
\keyword{datasets}
