% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{test_mejr_theme}
\alias{test_mejr_theme}
\title{Test theme by printing plots to pdf and viewport}
\usage{
test_mejr_theme(w = 6.875, h = 4.5, eplot = list(), mejr = list(),
  gg = theme(), print = TRUE, with_test_theme = FALSE, dir)
}
\arguments{
\item{w}{pdf width (inches)}

\item{h}{pdf height (inches)}

\item{eplot}{list of options for \code{example_plot}}

\item{mejr}{list of options for \code{theme_mejr(...)}}

\item{gg}{further theme customization with ggplot::theme(...)}

\item{print}{show the \code{eplot} plot}

\item{with_test_theme}{also print the same plot using \code{ggplot2::theme_test}}

\item{dir}{directory where to save temp pdf files, such as tempdir()}
}
\description{
Test theme by printing plots to pdf and viewport
}
