% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdistribute-package.R
\docType{package}
\name{ggdistribute-package}
\alias{ggdistribute-package}
\title{ggdistribute: \code{ggplot2} Extension for Plotting Distributions}
\description{
\code{ggdistribute} Uses the ggproto system to provide different geoms, stats, and
positions for plotting distributions.
}
\details{
\code{ggdistribute} ...
}
\seealso{
\link[ggplot2:geom_density]{ggplot2::geom_density}, \link[ggplot2:position_dodge]{ggplot2::position_dodge}

See \code{help(package = "ggdistribute")} for a list of functions.

View vignettes with \code{browseVignettes(package ="ggdistribute")}.
}
